tests/wpt/tests/css/css-ui/tentative/webkit-user-modify-02.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>-webkit-user-modify should be unsupported in contenteditable</title>
<link rel="help" href="https://developer.mozilla.org/en-US/docs/Web/CSS/user-modify">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="contentEditableTrue" contenteditable="true"></div>
<div id="contentEditablePlainText" contenteditable="plaintext-only"></div>
<script>
    test(function() {
        const contentEditableTrue = document.getElementById('contentEditableTrue');
        assert_equals(getComputedStyle(contentEditableTrue).getPropertyValue('-webkit-user-modify'), '');
    }, 'contenteditable="true" doesn\'t use unsupported -webkit-user-modify property');

    test(function() {
        const contentEditablePlainText = document.getElementById('contentEditablePlainText');
        assert_equals(getComputedStyle(contentEditablePlainText).getPropertyValue('-webkit-user-modify'), '');
    }, 'contenteditable="plaintext-only" doesn\'t use unsupported -webkit-user-modify property');
</script>
