third_party/blink/web_tests/external/wpt/css/css-ui/tentative/webkit-user-modify-02-expected.txt

This is a testharness.js-based test.
[FAIL] contenteditable="true" doesn't use unsupported -webkit-user-modify property
  assert_equals: expected "" but got "read-write"
[FAIL] contenteditable="plaintext-only" doesn't use unsupported -webkit-user-modify property
  assert_equals: expected "" but got "read-write-plaintext-only"
Harness: the test ran to completion.
